IDVA6-2037: Data migration skeleton. Updated and refactored data model. Updated models and interfaces, with light refactoring around globals and usage of request context.
