+++ src/main/java/uk/gov/companieshouse/accounts/association/models/AssociationDao.java
@@ -13,160 +13,256 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-@Document("user_company_associations")
-@CompoundIndex(name = "company_user_idx", def = "{'company_number': 1, 'user_id': 1, 'user_email': 1}", unique = true)
+@Document( "user_company_associations" )
+@CompoundIndex( name = "company_user_idx", def = "{'company_number': 1, 'user_id': 1, 'user_email': 1}", unique = true )
 public class AssociationDao {
+
     @Id
     private String id;
-    @NotNull
+
     @Indexed
-    @Field("company_number")
+    @NotNull
+    @Field( "company_number" )
     private String companyNumber;
 
-    @Field("user_id")
     @Indexed
+    @Field( "user_id" )
     private String userId;
+
+    @Indexed
+    @Field( "user_email" )
+    private String userEmail;
+
     @NotNull
     private String status;
-    @Field("created_at")
-    @CreatedDate
-    private LocalDateTime createdAt;
-    @Field("approved_at")
-    private LocalDateTime approvedAt;
-    @Field("removed_at")
-    private LocalDateTime removedAt;
+
     @NotNull
-    @Field("approval_route")
+    @Field( "approval_route" )
     private String approvalRoute;
-    @Field("user_email")
-    @Indexed
-    private String userEmail;
-    @Field("approval_expiry_at")
+
+    private final List<InvitationDao> invitations = new ArrayList<>();
+
+    @Field( "approval_expiry_at" )
     private LocalDateTime approvalExpiryAt;
 
-    @Field("invitations")
-    private final List<InvitationDao> invitations;
+    @Field( "approved_at" )
+    private LocalDateTime approvedAt;
+
+    @Field( "removed_at" )
+    private LocalDateTime removedAt;
+
+    @CreatedDate
+    @Field( "created_at" )
+    private LocalDateTime createdAt;
+
+    @Field( "migrated_at" )
+    private LocalDateTime migratedAt;
+
+    @Field( "previous_states" )
+    private final List<PreviousStatesDao> previousStates = new ArrayList<>();
+
     @NotNull
     private String etag;
+
     @Version
     private Integer version;
 
-    public AssociationDao() {
-        invitations = new ArrayList<>();
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
+    public void setId( final String id ){
         this.id = id;
     }
 
-    public String getCompanyNumber() {
-        return companyNumber;
+    public AssociationDao id( final String id ){
+        setId( id );
+        return this;
     }
 
-    public void setCompanyNumber(String companyNumber) {
+    public String getId(){
+        return id;
+    }
+
+    public void setCompanyNumber( final String companyNumber ){
         this.companyNumber = companyNumber;
     }
 
-    public String getUserId() {
-        return userId;
+    public AssociationDao companyNumber( final String companyNumber ){
+        setCompanyNumber( companyNumber );
+        return this;
     }
 
-    public void setUserId(String userId) {
+    public String getCompanyNumber(){
+        return companyNumber;
+    }
+
+    public void setUserId( final String userId ){
         this.userId = userId;
     }
 
-    public String getStatus() {
-        return status;
+    public AssociationDao userId( final String userId ){
+        setUserId( userId );
+        return this;
     }
 
-    public void setStatus(String status) {
-        this.status = status;
+    public String getUserId(){
+        return userId;
     }
 
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public String getApprovalRoute() {
-        return approvalRoute;
-    }
-
-    public void setApprovalRoute(String approvalRoute) {
-        this.approvalRoute = approvalRoute;
-    }
-
-    public String getUserEmail() {
-        return userEmail;
-    }
-
-    public void setUserEmail(String userEmail) {
+    public void setUserEmail( final String userEmail ){
         this.userEmail = userEmail;
     }
 
-    public List<InvitationDao> getInvitations() {
+    public AssociationDao userEmail( final String userEmail ){
+        setUserEmail( userEmail );
+        return this;
+    }
+
+    public String getUserEmail(){
+        return userEmail;
+    }
+
+    public void setStatus( final String status ){
+        this.status = status;
+    }
+
+    public AssociationDao status( final String status ){
+        setStatus( status );
+        return this;
+    }
+
+    public String getStatus(){
+        return status;
+    }
+
+    public void setApprovalRoute( final String approvalRoute ){
+        this.approvalRoute = approvalRoute;
+    }
+
+    public AssociationDao approvalRoute( final String approvalRoute ){
+        setApprovalRoute( approvalRoute );
+        return this;
+    }
+
+
+    public String getApprovalRoute(){
+        return approvalRoute;
+    }
+
+    public void setInvitations( final List<InvitationDao> invitations ){
+        this.invitations.clear();
+        this.invitations.addAll( invitations );
+    }
+
+    public AssociationDao invitations( final List<InvitationDao> invitations ){
+        setInvitations( invitations );
+        return this;
+    }
+
+    public List<InvitationDao> getInvitations(){
         return invitations;
     }
 
-    public void setInvitations(List<InvitationDao> invitations) {
-
-        this.invitations.addAll(invitations);
+    public void setApprovalExpiryAt( final LocalDateTime approvalExpiryAt ){
+        this.approvalExpiryAt = approvalExpiryAt;
     }
 
-    public String getEtag() {
-        return etag;
+    public AssociationDao approvalExpiryAt( final LocalDateTime approvalExpiryAt ){
+        setApprovalExpiryAt( approvalExpiryAt );
+        return this;
     }
 
+    public LocalDateTime getApprovalExpiryAt(){
+        return approvalExpiryAt;
+    }
 
-    public LocalDateTime getApprovedAt() {
+    public void setApprovedAt( final LocalDateTime approvedAt ){
+        this.approvedAt = approvedAt;
+    }
+
+    public AssociationDao approvedAt( final LocalDateTime approvedAt ){
+        setApprovedAt( approvedAt );
+        return this;
+    }
+
+    public LocalDateTime getApprovedAt(){
         return approvedAt;
     }
 
-    public void setApprovedAt(LocalDateTime approvedAt) {
-        this.approvedAt = approvedAt;
+    public void setRemovedAt( final LocalDateTime removedAt ){
+        this.removedAt = removedAt;
+    }
+
+    public AssociationDao removedAt( final LocalDateTime removedAt ){
+        setRemovedAt( removedAt );
+        return this;
     }
 
     public LocalDateTime getRemovedAt() {
         return removedAt;
     }
 
-    public void setRemovedAt(LocalDateTime removedAt) {
-        this.removedAt = removedAt;
+    public LocalDateTime getCreatedAt(){
+        return createdAt;
     }
 
-    public LocalDateTime getApprovalExpiryAt() {
-        return approvalExpiryAt;
+    public void setMigratedAt( final LocalDateTime migratedAt ){
+        this.migratedAt = migratedAt;
     }
 
-    public void setApprovalExpiryAt(LocalDateTime approvalExpiryAt) {
-        this.approvalExpiryAt = approvalExpiryAt;
+    public AssociationDao migratedAt( final LocalDateTime migratedAt ){
+        setMigratedAt( migratedAt );
+        return this;
     }
 
-    public void setEtag(String etag) {
+    public LocalDateTime getMigratedAt(){
+        return migratedAt;
+    }
+
+    public void setPreviousStates( final List<PreviousStatesDao> previousStates ){
+        this.previousStates.clear();
+        this.previousStates.addAll( previousStates );
+    }
+
+    public AssociationDao previousStates( final List<PreviousStatesDao> previousStates ){
+        setPreviousStates( previousStates );
+        return this;
+    }
+
+    public List<PreviousStatesDao> getPreviousStates(){
+        return previousStates;
+    }
+
+    public void setEtag( final String etag ){
         this.etag = etag;
     }
 
-    public Integer getVersion() {
+    public AssociationDao etag( final String etag ){
+        setEtag( etag );
+        return this;
+    }
+
+    public String getEtag(){
+        return etag;
+    }
+
+    public Integer getVersion(){
         return version;
     }
 
     @Override
     public String toString() {
-        return "Association{" +
+        return "AssociationDao{" +
                 "id='" + id + '\'' +
                 ", companyNumber='" + companyNumber + '\'' +
                 ", userId='" + userId + '\'' +
+                ", userEmail='" + userEmail + '\'' +
                 ", status='" + status + '\'' +
-                ", createdAt=" + createdAt +
+                ", approvalRoute='" + approvalRoute + '\'' +
+                ", invitations=" + invitations +
+                ", approvalExpiryAt=" + approvalExpiryAt +
                 ", approvedAt=" + approvedAt +
                 ", removedAt=" + removedAt +
-                ", approvalRoute='" + approvalRoute + '\'' +
-                ", userEmail='" + userEmail + '\'' +
-                ", approvalExpiryAt=" + approvalExpiryAt +
-                ", invitations=" + invitations +
+                ", createdAt=" + createdAt +
+                ", migratedAt=" + migratedAt +
+                ", previousStates=" + previousStates +
                 ", etag='" + etag + '\'' +
                 ", version=" + version +
                 '}';
