+++ src/main/java/uk/gov/companieshouse/accounts/association/models/Constants.java
@@ -5,6 +5,9 @@ public class Constants {
     public static final String X_REQUEST_ID = "X-Request-Id";
     public static final String UNKNOWN = "unknown";
     public static final String OAUTH2 = "oauth2";
+    public static final String PAGE_INDEX_WAS_LESS_THAN_0 = "pageIndex was less than 0";
+    public static final String PLEASE_CHECK_THE_REQUEST_AND_TRY_AGAIN = "Please check the request and try again";
+    public static final String ITEMS_PER_PAGE_WAS_LESS_THAN_OR_EQUAL_TO_0 = "itemsPerPage was less than or equal to 0";
 
 
 }
