+++ src/main/java/uk/gov/companieshouse/accounts/association/models/InvitationDao.java
@@ -7,29 +7,41 @@ import java.time.LocalDateTime;
 import static uk.gov.companieshouse.accounts.association.utils.StaticPropertyUtil.DAYS_SINCE_INVITE_TILL_EXPIRES;
 
 public class InvitationDao {
-    @Field("invited_by")
+
+    @Field( "invited_by" )
     private String invitedBy;
-    @Field("invited_at")
+
+    @Field( "invited_at" )
     private LocalDateTime invitedAt;
 
-    public String getInvitedBy() {
-        return invitedBy;
-    }
-
-    public void setInvitedBy(String invitedBy) {
+    public void setInvitedBy( final String invitedBy ){
         this.invitedBy = invitedBy;
     }
 
-    public LocalDateTime getInvitedAt() {
-        return invitedAt;
+    public InvitationDao invitedBy( final String invitedBy ){
+        setInvitedBy( invitedBy );
+        return this;
     }
 
-    public void setInvitedAt(LocalDateTime invitedAt) {
+    public String getInvitedBy(){
+        return invitedBy;
+    }
+
+    public void setInvitedAt( final LocalDateTime invitedAt ){
         this.invitedAt = invitedAt;
     }
 
+    public InvitationDao invitedAt( final LocalDateTime invitedAt ){
+        setInvitedAt( invitedAt );
+        return this;
+    }
+
+    public LocalDateTime getInvitedAt(){
+        return invitedAt;
+    }
+
     public LocalDateTime getExpiredAt() {
-        return invitedAt.plusDays(DAYS_SINCE_INVITE_TILL_EXPIRES);
+        return invitedAt.plusDays( DAYS_SINCE_INVITE_TILL_EXPIRES );
     }
 
     @Override
@@ -39,4 +51,5 @@ public class InvitationDao {
                 ", invitedAt=" + invitedAt +
                 '}';
     }
+
 }
