+++ src/test/java/uk/gov/companieshouse/accounts/association/controller/UserCompanyAssociationsTest.java
@@ -7,14 +7,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.mockito.internal.verification.Times;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers;
 import org.springframework.test.web.servlet.MockMvc;
@@ -25,7 +23,6 @@ import uk.gov.companieshouse.accounts.association.common.ComparisonUtils;
 import uk.gov.companieshouse.accounts.association.common.Mockers;
 import uk.gov.companieshouse.accounts.association.common.TestDataManager;
 import uk.gov.companieshouse.accounts.association.configuration.WebSecurityConfig;
-import uk.gov.companieshouse.accounts.association.exceptions.BadRequestRuntimeException;
 import uk.gov.companieshouse.accounts.association.exceptions.InternalServerErrorRuntimeException;
 import uk.gov.companieshouse.accounts.association.exceptions.NotFoundRuntimeException;
 import uk.gov.companieshouse.accounts.association.models.AssociationDao;
@@ -118,15 +115,6 @@ class UserCompanyAssociationsTest {
     }
 
     @Test
-    void fetchAssociationsByTestShouldThrow400ErrorWhenRequestIdNotProvided() throws Exception {
-        mockMvc.perform(get("/associations")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
     void fetchAssociationsByTestShouldReturnEmptyDataWhenNoAssociationsFoundForEricIdentity() throws Exception {
         final var user = testDataManager.fetchUserDtos( "000" ).getFirst();
 
@@ -199,26 +187,6 @@ class UserCompanyAssociationsTest {
         mockMvc.perform(get("/associations?items_per_page=0")
                         .header("X-Request-Id", "theId123")
                         .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void getInvitationsForAssociationWithUnacceptablePageIndexReturnsBadRequest() throws Exception {
-        mockMvc.perform(get("/associations/1/invitations?page_index=-1&items_per_page=1")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void getInvitationsForAssociationWithUnacceptableItemsPerPageReturnsBadRequest() throws Exception {
-        mockMvc.perform(get("/associations/1/invitations?page_index=0&items_per_page=-1")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
                         .header("ERIC-Identity-Type", "oauth2")
                         .header("ERIC-Authorised-Key-Roles", "*"))
                 .andExpect(status().isBadRequest());
@@ -445,72 +413,30 @@ class UserCompanyAssociationsTest {
     }
 
     @Test
-    void getAssociationDetailsWithoutPathVariableReturnsNotFound() throws Exception {
-        mockMvc.perform(get("/associations/")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isNotFound());
-    }
+    void fetchAssociationsByCanFetchMigratedAssociation() throws Exception {
+        final var user = testDataManager.fetchUserDtos( "MKUser001" ).getFirst();
+        final var expectedAssociationsList = new AssociationsList()
+                .items( List.of( testDataManager.fetchAssociationDto( "MKAssociation001", user  ) ) )
+                .links( new Links().self( "/associations?page_index=0&items_per_page=15" ).next( "" ) )
+                .itemsPerPage( 15 ).pageNumber( 0 ).totalPages( 1 ).totalResults( 1 );
 
-    @Test
-    void getAssociationsDetailsWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(get("/associations/1")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
+        mockers.mockUsersServiceFetchUserDetails( "MKUser001" );
+        Mockito.doReturn( expectedAssociationsList ).when( associationsService ).fetchAssociationsForUserStatusAndCompany( user, List.of( StatusEnum.MIGRATED.getValue() ), 0, 15, null );
 
-    @Test
-    void getAssociationDetailsWithMalformedInputReturnsBadRequest() throws Exception {
-        mockMvc.perform(get("/associations/$")
-                .header("X-Request-Id", "theId123")
-                .header("Eric-identity", "000")
-                .header("ERIC-Identity-Type", "oauth2")
-                .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
+        final var response =
+                mockMvc.perform( get( "/associations?status=migrated" )
+                                .header( "X-Request-Id", "theId123" )
+                                .header( "Eric-identity", "MKUser001" )
+                                .header( "ERIC-Identity-Type", "oauth2" ) )
+                        .andExpect( status().isOk() );
 
-    @Test
-    void getAssociationUserDetailsWithNonexistentIdReturnsNotFound() throws Exception {
-        Mockito.doReturn(Optional.empty()).when(associationsService).findAssociationById("11");
-        mockMvc.perform(get("/associations/11")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isNotFound());
-    }
+        final var associationsList = parseResponseTo( response, AssociationsList.class );
+        final var associations = associationsList.getItems();
+        final var associationOne = associations.getFirst();
 
-    @Test
-    void getAssociationDetailsFetchesAssociationDetails() throws Exception {
-        final var user = testDataManager.fetchUserDtos( "9999" ).getFirst();
-        final var association = testDataManager.fetchAssociationDto( "18", user );
-
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationById("18");
-
-        final var response = mockMvc.perform(get("/associations/18")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isOk());
-        final var result = parseResponseTo( response, Association.class );
-
-        Assertions.assertEquals("18", result.getId());
-    }
-
-    @Test
-    void addAssociationWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"111111\"}"))
-                .andExpect(status().isBadRequest());
+        Assertions.assertEquals( "MKAssociation001", associationOne.getId() );
+        Assertions.assertEquals( "migrated", associationOne.getStatus().getValue() );
+        Assertions.assertEquals( "migration", associationOne.getApprovalRoute().getValue() );
     }
 
     @Test
@@ -603,7 +529,7 @@ class UserCompanyAssociationsTest {
                 .andExpect(status().isCreated());
         final var responseBodyPost = parseResponseTo( response, ResponseBodyPost.class );
 
-        Assertions.assertEquals("1", responseBodyPost.getAssociationId());
+        Assertions.assertEquals("/associations/1", responseBodyPost.getAssociationLink());
     }
 
     @Test
@@ -716,953 +642,6 @@ class UserCompanyAssociationsTest {
                 .andExpect( status().isCreated() );
 
         Mockito.verify( emailService ).sendAuthCodeConfirmationEmailToAssociatedUser( eq( "theId123" ), argThat( comparisonUtils.compare( company, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq( "Scrooge McDuck" ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(patch("/associations/18")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"removed\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithMalformedAssociationIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(patch("/associations/$$$")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"removed\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithoutRequestBodyReturnsBadRequest() throws Exception {
-        mockMvc.perform(patch("/associations/18")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithoutStatusReturnsBadRequest() throws Exception {
-        mockMvc.perform(patch("/associations/18")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithMalformedStatusReturnsBadRequest() throws Exception {
-        mockMvc.perform(patch("/associations/18")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"complicated\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNonexistentAssociationIdReturnsNotFound() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        Mockito.doReturn(Optional.empty()).when(associationsService).findAssociationById("9191");
-
-        mockMvc.perform(patch("/associations/9191")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"confirmed\"}"))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithRemovedUpdatesAssociationStatus() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "35" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "000" );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-        Mockito.doReturn( Optional.of( association ) ).when( associationsService ).findAssociationDaoById( "35" );
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendAuthorisationRemovedEmailToRemovedUser( eq( "theId123" ), eq( "333333" ), any( Mono.class ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendAuthorisationRemovedEmailToAssociatedUser( eq( "theId123" ), eq( "333333" ), any( Mono.class ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(patch("/associations/35")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"removed\"}"))
-                .andExpect(status().isOk());
-
-        Mockito.verify(associationsService).updateAssociation(anyString(), any(Update.class));
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithConfirmedUpdatesWithNoUserFoundShouldThrow404AssociationStatus() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        Mockito.doReturn( Optional.of( association ) ).when( associationsService ).findAssociationDaoById( "36" );
-        Mockito.doReturn( null ).when(usersService).searchUserDetails(List.of("light.yagami@death.note"));
-
-        mockMvc.perform(patch("/associations/36")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"confirmed\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndExistingUserAndConfirmedUpdatesAssociationStatus() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockUsersServiceFetchUserDetails( "000" );
-        Mockito.doReturn( Optional.of( association ) ).when( associationsService ).findAssociationDaoById( "36" );
-
-        mockMvc.perform(patch("/associations/36")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"confirmed\"}"))
-                .andExpect(status().isOk());
-
-        Mockito.verify(associationsService, new Times(1)).updateAssociation(anyString(), any(Update.class));
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndNonexistentUserAndConfirmedReturnsBadRequest() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockUsersServiceFetchUserDetails( "000" );
-        Mockito.doReturn( Optional.of( association ) ).when( associationsService ).findAssociationDaoById( "36" );
-
-        mockMvc.perform(patch("/associations/36")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"confirmed\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndExistingUserAndRemovedUpdatesAssociationStatus() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "34" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "000" );
-        Mockito.doReturn( Optional.of( association ) ).when( associationsService ).findAssociationDaoById( "34" );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationRejectedEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(patch("/associations/34")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"removed\"}"))
-                .andExpect(status().isOk());
-
-        Mockito.verify(associationsService, new Times(1)).updateAssociation(anyString(), any(Update.class));
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndNonexistentUserAndRemovedUpdatesAssociationStatus() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "34" ).getFirst();
-
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockUsersServiceFetchUserDetails( "000" );
-        Mockito.doReturn( Optional.of( association ) ).when( associationsService ).findAssociationDaoById( "34" );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationRejectedEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(patch("/associations/34")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"status\":\"removed\"}"))
-                .andExpect(status().isOk());
-
-        Mockito.verify(associationsService, new Times(1)).updateAssociation(anyString(), any(Update.class));
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsWhereTargetUserIdExistsSendsNotification() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "35" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999", "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("35");
-        Mockito.doReturn( List.of( "9999" ) ).when( associationsService ).fetchAssociatedUsers( "333333" );
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendAuthorisationRemovedEmailToRemovedUser( eq( "theId123" ), eq( "333333" ), any( Mono.class ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendAuthorisationRemovedEmailToAssociatedUser( eq( "theId123" ), eq( "333333" ), any( Mono.class ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform( patch( "/associations/35" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendAuthorisationRemovedEmailToAssociatedUser( eq( "theId123" ), eq( "333333" ), any( Mono.class ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsWhereTargetUserIdDoesNotExistSendsNotification() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "34" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "111", "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("34");
-        Mockito.doReturn( List.of( "111" ) ).when( associationsService ).fetchAssociatedUsers( "111111" );
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteCancelledEmail( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Batman" ), eq( association ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationCancelledEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Batman" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform( patch( "/associations/34" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendInvitationCancelledEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Batman" ), eq( "light.yagami@death.note" ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsWhereUserCannotBeFoundSendsNotification() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "34" ).getFirst();
-
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockUsersServiceFetchUserDetails( "111" );
-        mockers.mockCompanyServiceFetchCompanyProfile("111111");
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("34");
-        Mockito.doReturn( List.of("000" ) ).when( associationsService ).fetchAssociatedUsers( "111111" );
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteCancelledEmail( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Batman" ), eq( association ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationCancelledEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Batman" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform( patch( "/associations/34" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendInvitationCancelledEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Batman" ), eq( "light.yagami@death.note" ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsUsesDisplayNamesWhenAvailable() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "4" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "333" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-        mockers.mockUsersServiceSearchUserDetails( "444" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("4");
-        Mockito.doReturn( List.of( "333" ) ).when( associationsService ).fetchAssociatedUsers( "111111" );
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendAuthorisationRemovedEmailToRemovedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Harleen Quinzel" ), eq( "Boy Wonder" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendAuthorisationRemovedEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Harleen Quinzel" ), eq( "Boy Wonder" ) );
-
-        mockMvc.perform( patch( "/associations/4" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "333")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendAuthorisationRemovedEmailToRemovedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Harleen Quinzel" ), eq( "444" ) );
-        Mockito.verify( emailService ).sendAuthorisationRemovedEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), eq( "Harleen Quinzel" ), eq("Boy Wonder") );
-    }
-
-    @Test
-    void updateAssociationStatusForIdUserAcceptedInvitationNotificationsSendsNotification() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "6" ).getFirst();
-        final var companyDetails = testDataManager.fetchCompanyDetailsDtos( "111111" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "666", "222" );
-        mockers.mockUsersServiceSearchUserDetails( "666" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("6");
-        Mockito.doReturn( List.of( "222", "444" ) ).when( associationsService ).fetchAssociatedUsers( "111111" );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationAcceptedEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), any( Mono.class ), eq( "homer.simpson@springfield.com" ) );
-
-        mockMvc.perform( patch( "/associations/6" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "666")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendInvitationAcceptedEmailToAssociatedUser( eq( "theId123" ), eq( "111111" ), any( Mono.class ), any( Mono.class ), eq( "homer.simpson@springfield.com" ) );
-    }
-
-    @Test
-    void confirmUserStatusForExistingAssociationWithoutOneLoginUserShouldThrow400BadRequest() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "6" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "666", "222" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "homer.simpson@springfield.com" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("6");
-        Mockito.doReturn( List.of( "222", "444" ) ).when( associationsService ).fetchAssociatedUsers( "111111" );
-
-        mockMvc.perform( patch( "/associations/6" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "666")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void confirmUserStatusRequestForExistingAssociationWithoutOneLoginUserAndDifferentRequestingUserShouldThrow400BadRequest() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "6" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "666", "222" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "homer.simpson@springfield.com" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("6");
-        Mockito.doReturn( List.of( "222", "444" ) ).when( associationsService ).fetchAssociatedUsers( "111111" );
-
-        mockMvc.perform( patch( "/associations/6" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "222")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdUserAcceptedInvitationNotificationsUsesDisplayNamesWhenAvailable() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "38" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999", "444" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "scrooge.mcduck@disney.land" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "x222222" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("38");
-        Mockito.doReturn( List.of( "111", "222", "444" ) ).when( associationsService ).fetchAssociatedUsers( "x222222" );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationAcceptedEmailToAssociatedUser( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), any( Mono.class ), eq( "Scrooge McDuck" ) );
-
-        mockMvc.perform( patch( "/associations/38" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendInvitationAcceptedEmailToAssociatedUser( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), any(), eq( "Scrooge McDuck" ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdUserCancelledInvitationNotificationsSendsNotification() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "38" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999", "222" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "scrooge.mcduck@disney.land" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "x222222" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("38");
-        Mockito.doReturn( List.of( "111", "222", "444" ) ).when( associationsService ).fetchAssociatedUsers( "x222222" );
-        Mockito.doReturn( true ).when( associationsService ).confirmedAssociationExists( "x222222", "222" );
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteCancelledEmail( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), eq( "the.joker@gotham.city" ), eq( association ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationCancelledEmailToAssociatedUser( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), eq( "the.joker@gotham.city" ), eq( "Scrooge McDuck" ) );
-
-        mockMvc.perform( patch( "/associations/38" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "222")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendInvitationCancelledEmailToAssociatedUser( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), eq( "the.joker@gotham.city" ), eq( "Scrooge McDuck" ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdUserRejectedInvitationNotificationsSendsNotification() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "38" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "scrooge.mcduck@disney.land" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "x222222" );
-        Mockito.doReturn(Optional.of(association)).when(associationsService).findAssociationDaoById("38");
-        Mockito.doReturn( List.of( "111", "222", "444" ) ).when( associationsService ).fetchAssociatedUsers( "x222222" );
-
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationRejectedEmailToAssociatedUser( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), eq( "Scrooge McDuck" ) );
-
-        mockMvc.perform( patch( "/associations/38" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( emailService ).sendInvitationRejectedEmailToAssociatedUser( eq( "theId123" ), eq( "x222222" ), any( Mono.class ), eq( "Scrooge McDuck" ) );
-    }
-
-    @Test
-    void inviteUserWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations/invitations")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWithMalformedEricIdentityReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "$$$$" );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "$$$$")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}"))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void inviteUserWithNonexistentEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "9191" );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9191")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}"))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void inviteUserWithoutRequestBodyReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWithoutCompanyNumberReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"invitee_email_id\":\"bruce.wayne@gotham.city\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWithMalformedCompanyNumberReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"$$$$$$\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWithNonexistentCompanyNumberReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfileNotFound( "919191" );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"919191\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWithoutInviteeEmailIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWithMalformedInviteeEmailIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\",\"invitee_email_id\":\"$$$\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeEmailAndCompanyNumberExistsAndInviteeUserIsFoundPerformsSwapAndUpdateOperations() throws Exception {
-        final var associationDao = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-        final var associationDaoForComparison = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-        final var companyDetails = testDataManager.fetchCompanyDetailsDtos( "444444" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        Mockito.doReturn(Optional.of(associationDao)).when(associationsService).fetchAssociationForCompanyNumberAndUserEmail("444444", "light.yagami@death.note");
-        Mockito.doReturn(associationDao).when(associationsService).sendNewInvitation(eq("9999"), any());
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteEmail( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), any( String.class ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationEmailToAssociatedUser( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}"))
-                .andExpect(status().isCreated());
-
-        Mockito.verify(associationsService).sendNewInvitation(eq("9999"), argThat( comparisonUtils.compare( associationDaoForComparison, List.of( "id", "companyNumber", "createdAt", "approvedAt", "removedAt", "approvalRoute", "approvalExpiryAt", "invitations", "etag", "version" ), List.of( "userId", "userEmail" ), Map.of() ) ));
-        Mockito.verify( emailService ).sendInviteEmail( eq( "theId123" ), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), anyString(), eq( "light.yagami@death.note" ) );
-        Mockito.verify( emailService ).sendInvitationEmailToAssociatedUser( eq( "theId123" ), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), eq("light.yagami@death.note") );
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeEmailAndCompanyNumberExistsAndInviteeUserIsNotFoundDoesNotPerformSwapButDoesPerformUpdateOperation() throws Exception {
-        final var associationDao = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-        final var associationDaoForComparison = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-        final var companyDetails = testDataManager.fetchCompanyDetailsDtos( "444444" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "000" );
-        Mockito.doReturn(Optional.of(associationDao)).when(associationsService).fetchAssociationForCompanyNumberAndUserEmail("444444", "light.yagami@death.note");
-        Mockito.doReturn(associationDao).when(associationsService).sendNewInvitation(eq("9999"), any());
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteEmail( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), any( String.class ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationEmailToAssociatedUser( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}"))
-                .andExpect(status().isCreated());
-
-        Mockito.verify(associationsService).sendNewInvitation(eq("9999"), argThat(comparisonUtils.compare( associationDaoForComparison, List.of( "id", "companyNumber", "createdAt", "approvedAt", "removedAt", "approvalRoute", "approvalExpiryAt", "invitations", "etag", "version", "userId", "userEmail" ), List.of(), Map.of() )));
-        Mockito.verify( emailService ).sendInviteEmail( eq( "theId123" ), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), anyString(), eq( "light.yagami@death.note" ) );
-        Mockito.verify( emailService ).sendInvitationEmailToAssociatedUser(  eq("theId123"), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), eq("light.yagami@death.note") );
-    }
-
-    @Test
-    void inviteUserWhereInviteeUserIsFoundAndAssociationBetweenInviteeUserIdAndCompanyNumberExistsDoesNotPerformSwapButDoesPerformUpdateOperation() throws Exception {
-        final var requestingUserAssociation = testDataManager.fetchAssociationDaos( "19" ).getFirst();
-        final var targetUserAssociation = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-        final var companyDetails = testDataManager.fetchCompanyDetailsDtos( "444444" ).getFirst();
-
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        Mockito.doReturn(Optional.empty()).when(associationsService).fetchAssociationForCompanyNumberAndUserEmail("444444", "light.yagami@death.note");
-        Mockito.doReturn(Optional.of(targetUserAssociation)).when(associationsService).fetchAssociationForCompanyNumberAndUserId("444444", "000");
-        Mockito.doReturn(requestingUserAssociation).when(associationsService).sendNewInvitation(eq("9999"), any());
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteEmail( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), any( String.class ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationEmailToAssociatedUser( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}"))
-                .andExpect(status().isCreated());
-
-        Mockito.verify(associationsService).sendNewInvitation(eq("9999"), any() );
-        Mockito.verify( emailService ).sendInviteEmail( eq( "theId123" ), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), anyString(), eq( "light.yagami@death.note" ) );
-        Mockito.verify( emailService ).sendInvitationEmailToAssociatedUser( eq("theId123"), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), eq("light.yagami@death.note") );
-    }
-
-    @Test
-    void inviteUserWhereInviteeUserIsFoundAndAssociationBetweenInviteeUserIdAndCompanyNumberDoesNotExistCreatesNewAssociation() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "19" ).getFirst();
-        final var companyDetails = testDataManager.fetchCompanyDetailsDtos( "444444" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        Mockito.doReturn(Optional.empty()).when(associationsService).fetchAssociationForCompanyNumberAndUserEmail("444444", "light.yagami@death.note");
-        Mockito.doReturn(Optional.empty()).when(associationsService).fetchAssociationForCompanyNumberAndUserId("444444", "000");
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( association).when( associationsService ).createAssociation( "444444", "000", null, ApprovalRouteEnum.INVITATION, "9999" );
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteEmail( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), any( String.class ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationEmailToAssociatedUser( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}"))
-                .andExpect(status().isCreated());
-
-        Mockito.verify( emailService ).sendInviteEmail( eq( "theId123" ), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), anyString(), eq( "light.yagami@death.note" ) );
-        Mockito.verify( emailService ).sendInvitationEmailToAssociatedUser( eq("theId123"), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), eq("light.yagami@death.note") );
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeUserEmailAndCompanyNumberDoesNotExistAndInviteeUserIsNotFoundCreatesNewAssociation() throws Exception {
-        final var newAssociation = testDataManager.fetchAssociationDaos( "36" ).getFirst();
-        final var companyDetails = testDataManager.fetchCompanyDetailsDtos( "444444" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        Mockito.doReturn(Optional.empty()).when(associationsService).fetchAssociationForCompanyNumberAndUserEmail("444444", "light.yagami@death.note");
-        Mockito.doReturn(Optional.empty()).when(associationsService).fetchAssociationForCompanyNumberAndUserId("444444", "000");
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-        Mockito.doReturn( newAssociation).when( associationsService ).createAssociation( "444444", null, "light.yagami@death.note", ApprovalRouteEnum.INVITATION, "9999" );
-        Mockito.doReturn( Mono.empty() ).when( emailService ).sendInviteEmail( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), any( String.class ), eq( "light.yagami@death.note" ) );
-        Mockito.doReturn( sendEmailMock ).when( emailService ).sendInvitationEmailToAssociatedUser( eq( "theId123" ), eq( companyDetails ), eq( "Scrooge McDuck" ), eq( "light.yagami@death.note" ) );
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}"))
-                .andExpect(status().isCreated());
-
-        Mockito.verify( emailService ).sendInviteEmail( eq( "theId123" ), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), anyString(), eq( "light.yagami@death.note" ) );
-        Mockito.verify( emailService ).sendInvitationEmailToAssociatedUser( eq("theId123"), argThat( comparisonUtils.compare( companyDetails, List.of( "companyNumber", "companyName" ), List.of(), Map.of() ) ), eq("Scrooge McDuck"), eq("light.yagami@death.note") );
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeUserEmailAndCompanyNumberExistAndInviteeUserIsFoundThrowsBadRequest() throws Exception {
-        final var associationDao = testDataManager.fetchAssociationDaos( "35" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        Mockito.doReturn(Optional.empty()).when(associationsService).fetchAssociationForCompanyNumberAndUserEmail("333333", "light.yagami@death.note");
-        Mockito.doReturn(Optional.of(associationDao)).when(associationsService).fetchAssociationForCompanyNumberAndUserId("333333", "000");
-        Mockito.doThrow(new BadRequestRuntimeException("There is an existing association with Confirmed status for the user")).when(associationsService).fetchAssociationForCompanyNumberAndUserId("333333", "000");
-        Mockito.when(associationsService.confirmedAssociationExists(Mockito.any(), Mockito.any())).thenReturn(true);
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\",\"invitee_email_id\":\"light.yagami@death.note\"}"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void getInvitationsForAssociationWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform(get("/associations/1/invitations")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void getInvitationsForAssociationWithMalformedInputReturnsBadRequest() throws Exception {
-        mockMvc.perform(get("/associations/$/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void getInvitationsForAssociationWithNonexistentIdReturnsNotFound() throws Exception {
-        Mockito.doReturn(Optional.empty()).when(associationsService).findAssociationDaoById("11");
-
-        mockMvc.perform(get("/associations/11/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isNotFound());
-    }
-
-    @Test
-    void getInvitationsForAssociationFetchesInvitations() throws Exception {
-        final var associationDao = testDataManager.fetchAssociationDaos( "37" ).getFirst();
-        final var invitations = testDataManager.fetchInvitations( "37" );
-
-        Mockito.doReturn(Optional.of(associationDao)).when(associationsService).findAssociationDaoById("37");
-        Mockito.doReturn( new InvitationsList().items( invitations ) ).when(associationsService).fetchInvitations(associationDao, 0, 15);
-
-        final var response = mockMvc.perform(get("/associations/37/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isOk());
-        final var resultInvitationsList = parseResponseTo( response, InvitationsList.class );
-        final var resultInvitations = resultInvitationsList.getItems();
-
-        Assertions.assertEquals( 3, resultInvitations.size() );
-        Assertions.assertEquals( resultInvitations.getFirst(), invitations.getFirst() );
-        Assertions.assertEquals( resultInvitations.get(1), invitations.get(1) );
-        Assertions.assertEquals( resultInvitations.getLast(), invitations.getLast() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithoutEricIdentityReturnsForbidden() throws Exception {
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithMalformedEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "$$$$" );
-
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "$$$$")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithNonexistentEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "9191" );
-
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9191")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithUnacceptablePageIndexReturnsBadRequest() throws Exception {
-        mockMvc.perform( get( "/associations/invitations?page_index=-1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithUnacceptableItemsPerPageReturnsBadRequest() throws Exception {
-        mockMvc.perform( get( "/associations/invitations?page_index=0&items_per_page=-1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserRetrievesActiveInvitationsInCorrectOrderAndPaginatesCorrectly() throws Exception {
-        final var user = testDataManager.fetchUserDtos( "9999" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                            .header("X-Request-Id", "theId123")
-                            .header("Eric-identity", "9999")
-                            .header("ERIC-Identity-Type", "oauth2")
-                            .header("ERIC-Authorised-Key-Roles", "*") )
-                    .andExpect( status().isOk() );
-
-        Mockito.verify( associationsService ).fetchActiveInvitations( argThat( comparisonUtils.compare( user, List.of( "userId" ), List.of(), Map.of() ) ), eq( 1 ), eq( 1 ) );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithoutPageIndexAndItemsPerPageUsesDefaults() throws Exception {
-        final var user = testDataManager.fetchUserDtos( "9999" ).getFirst();
-
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( get( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isOk() );
-
-        Mockito.verify( associationsService ).fetchActiveInvitations( argThat( comparisonUtils.compare( user, List.of( "userId" ), List.of(), Map.of() ) ), eq( 0 ), eq( 15 ) );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithPaginationAndVerifyResponse() throws Exception {
-        final var user = testDataManager.fetchUserDtos( "000" ).getFirst();
-        final var invitations = testDataManager.fetchInvitations( "37" );
-
-        final var mockLinks = new Links()
-                .self("/associations/invitations?page_index=1&items_per_page=1")
-                .next("/associations/invitations?page_index=2&items_per_page=1");
-
-        final var mockInvitationsList = new InvitationsList()
-                .items(invitations).links(mockLinks)
-                .itemsPerPage(1).pageNumber(1).totalResults(2).totalPages(2);
-
-        mockers.mockUsersServiceFetchUserDetails( "000" );
-        when(associationsService.fetchActiveInvitations(user, 1,1)).thenReturn(mockInvitationsList);
-
-        final var response = mockMvc.perform(get("/associations/invitations?page_index=1&items_per_page=1")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "000")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isOk());
-
-        Mockito.verify(associationsService).fetchActiveInvitations( user,1,1);
-
-        final var invitationsList = parseResponseTo( response, InvitationsList.class );
-
-        assertNotNull(invitationsList);
-        assertEquals(1, invitationsList.getItemsPerPage().intValue());
-        assertEquals(1, invitationsList.getPageNumber().intValue());
-        assertEquals(2, invitationsList.getTotalResults().intValue());
-        assertEquals(2, invitationsList.getTotalPages().intValue());
-        assertNotNull(invitationsList.getLinks());
-        assertEquals("/associations/invitations?page_index=1&items_per_page=1", invitationsList.getLinks().getSelf());
-        assertEquals("/associations/invitations?page_index=2&items_per_page=1", invitationsList.getLinks().getNext());
-
-    }
-
-    @Test
-    void getInvitationsForAssociationWithPaginationAndVerifyResponse() throws Exception {
-        final var association = testDataManager.fetchAssociationDaos( "37" ).getFirst();
-        final var invitations = testDataManager.fetchInvitations( "37" );
-
-        final var mockLinks = new Links()
-                .self("/associations/37/invitations?page_index=1&items_per_page=1")
-                .next("/associations/37/invitations?page_index=2&items_per_page=1");
-
-        final var mockInvitationsList = new InvitationsList()
-                .items(invitations).links(mockLinks)
-                .itemsPerPage(1).pageNumber(1).totalResults(2).totalPages(2);
-
-        when(associationsService.findAssociationDaoById("37")).thenReturn(Optional.of(association));
-        when(associationsService.fetchInvitations(association, 1, 1)).thenReturn(mockInvitationsList);
-
-        final var response = mockMvc.perform(get("/associations/37/invitations?page_index=1&items_per_page=1")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*"))
-                .andExpect(status().isOk());
-        final var invitationsList = parseResponseTo( response, InvitationsList.class );
-
-        Mockito.verify(associationsService).findAssociationDaoById("37");
-        Mockito.verify(associationsService).fetchInvitations(association, 1, 1);
-
-        assertNotNull(invitationsList);
-        assertEquals(1, invitationsList.getItemsPerPage().intValue());
-        assertEquals(1, invitationsList.getPageNumber().intValue());
-        assertEquals(2, invitationsList.getTotalResults().intValue());
-        assertEquals(2, invitationsList.getTotalPages().intValue());
-        assertNotNull(invitationsList.getLinks());
-        assertEquals("/associations/37/invitations?page_index=1&items_per_page=1", invitationsList.getLinks().getSelf());
-        assertEquals("/associations/37/invitations?page_index=2&items_per_page=1", invitationsList.getLinks().getNext());
-
-    }
-
-    @Test
-    void whenConfirmedAssociationDoesNotExist_thenThrowsBadRequestException() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        when(associationsService.confirmedAssociationExists(anyString(), anyString())).thenReturn(false);
-
-        mockMvc.perform(post("/associations/invitations")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{\"company_number\":\"333333\",\"invitee_email_id\":\"russell.howard@comedy.com\"}"))
-                .andExpect(status().isBadRequest());
     }
 
 }
