+++ src/test/java/uk/gov/companieshouse/accounts/association/integration/AssociationsListForCompanyControllerTest.java
@@ -95,14 +95,6 @@ class AssociationsListForCompanyControllerTest {
     }
 
     @Test
-    void getAssociationsForCompanyWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockMvc.perform( get( "/associations/companies/111111" )
-                .header( "ERIC-Identity", "111" )
-                .header( "ERIC-Identity-Type", "oauth2" ) )
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
     void getAssociationsForCompanyWithoutQueryParamsUsesDefaults() throws Exception {
         associationsRepository.insert( testDataManager.fetchAssociationDaos( "1", "2", "3", "4", "5", "6", "7", "8", "9", "10","11","12","13","14","15","16","17" ) );
         mockers.mockUsersServiceFetchUserDetails( "111", "222", "333", "444", "555", "666", "777", "888", "999" ,"1111", "2222", "3333", "4444", "5555", "6666", "7777" );
