+++ src/test/java/uk/gov/companieshouse/accounts/association/integration/UserCompanyAssociationTest.java
@@ -1,0 +1,492 @@
+package uk.gov.companieshouse.accounts.association.integration;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.times;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static uk.gov.companieshouse.accounts.association.common.ParsingUtils.localDateTimeToNormalisedString;
+import static uk.gov.companieshouse.accounts.association.common.ParsingUtils.parseResponseTo;
+import static uk.gov.companieshouse.accounts.association.utils.MessageType.INVITATION_ACCEPTED_MESSAGE_TYPE;
+import static uk.gov.companieshouse.accounts.association.utils.MessageType.INVITE_CANCELLED_MESSAGE_TYPE;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.format.annotation.DateTimeFormat.ISO;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import uk.gov.companieshouse.accounts.association.common.ComparisonUtils;
+import uk.gov.companieshouse.accounts.association.common.Mockers;
+import uk.gov.companieshouse.accounts.association.common.TestDataManager;
+import uk.gov.companieshouse.accounts.association.models.AssociationDao;
+import uk.gov.companieshouse.accounts.association.models.email.builders.InvitationAcceptedEmailBuilder;
+import uk.gov.companieshouse.accounts.association.repositories.AssociationsRepository;
+import uk.gov.companieshouse.accounts.association.service.CompanyService;
+import uk.gov.companieshouse.accounts.association.service.UsersService;
+import uk.gov.companieshouse.api.accounts.associations.model.Association;
+import uk.gov.companieshouse.api.accounts.associations.model.Association.ApprovalRouteEnum;
+import uk.gov.companieshouse.api.accounts.associations.model.Association.StatusEnum;
+import uk.gov.companieshouse.api.accounts.associations.model.AssociationLinks;
+import uk.gov.companieshouse.api.accounts.associations.model.RequestBodyPut;
+import uk.gov.companieshouse.api.handler.exception.URIValidationException;
+import uk.gov.companieshouse.email_producer.EmailProducer;
+import uk.gov.companieshouse.email_producer.factory.KafkaProducerFactory;
+
+@AutoConfigureMockMvc
+@SpringBootTest
+@ExtendWith( MockitoExtension.class )
+@Tag( "integration-test" )
+class UserCompanyAssociationTest {
+
+    @Autowired
+    private MongoTemplate mongoTemplate;
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockBean
+    private CompanyService companyService;
+
+    @MockBean
+    private UsersService usersService;
+
+    @MockBean
+    private EmailProducer emailProducer;
+
+    @MockBean
+    private KafkaProducerFactory kafkaProducerFactory;
+
+    @Autowired
+    private AssociationsRepository associationsRepository;
+
+    private TestDataManager testDataManager = TestDataManager.getInstance();
+
+    private CountDownLatch latch;
+
+    private Mockers mockers;
+
+    private ComparisonUtils comparisonUtils = new ComparisonUtils();
+
+    private void setEmailProducerCountDownLatch( int countdown ){
+        latch = new CountDownLatch( countdown );
+        doAnswer( invocation -> {
+            latch.countDown();
+            return null;
+        } ).when( emailProducer ).sendEmail( any(), any() );
+    }
+
+    @BeforeEach
+    public void setup() throws IOException, URIValidationException {
+        mockers = new Mockers( null, emailProducer, companyService, usersService );
+    }
+
+    @Test
+    void getAssociationDetailsWithoutPathVariableReturnsNotFound() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails( "9999" );
+
+        mockMvc.perform( get( "/associations/" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*") )
+                .andExpect( status().isNotFound() );
+    }
+
+    @Test
+    void getAssociationUserDetailsWithNonexistentIdReturnsNotFound() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails( "9999" );
+
+        mockMvc.perform( get( "/associations/1" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*") )
+                .andExpect( status().isNotFound() );
+    }
+
+
+    @Test
+    void getAssociationDetailsFetchesAssociationDetails() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "18" ) );
+        mockers.mockUsersServiceFetchUserDetails( "9999" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
+
+        final var response =
+                mockMvc.perform( get( "/associations/18" )
+                                .header("X-Request-Id", "theId123")
+                                .header("Eric-identity", "9999")
+                                .header("ERIC-Identity-Type", "oauth2")
+                                .header("ERIC-Authorised-Key-Roles", "*") )
+                        .andExpect( status().isOk() );
+
+        final var association = parseResponseTo( response, Association.class );
+
+        Assertions.assertEquals( "9999", association.getUserId());
+    }
+
+    @Test
+    void getAssociationForIdCanFetchMigratedAssociation() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "MKAssociation001" ) );
+        mockers.mockUsersServiceFetchUserDetails( "MKUser001" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "MKCOMP001" );
+
+        final var response =
+                mockMvc.perform( get( "/associations/MKAssociation001" )
+                                .header( "X-Request-Id", "theId123" )
+                                .header( "Eric-identity", "MKUser001" )
+                                .header( "ERIC-Identity-Type", "oauth2" ) )
+                        .andExpect( status().isOk() );
+
+        final var association = parseResponseTo( response, Association.class );
+
+        Assertions.assertEquals( "MKAssociation001", association.getId() );
+        Assertions.assertEquals( "migrated", association.getStatus().getValue() );
+        Assertions.assertEquals( "migration", association.getApprovalRoute().getValue() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithMalformedAssociationIdReturnsBadRequest() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+
+        mockMvc.perform( patch( "/associations/$$$" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isBadRequest() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithoutRequestBodyReturnsBadRequest() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+
+        mockMvc.perform( patch( "/associations/18" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON ) )
+                .andExpect( status().isBadRequest() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithoutStatusReturnsBadRequest() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+
+        mockMvc.perform( patch( "/associations/18" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{}" ) )
+                .andExpect( status().isBadRequest() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithMalformedStatusReturnsBadRequest() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+
+        mockMvc.perform( patch( "/associations/18")
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"complicated\"}" ) )
+                .andExpect( status().isBadRequest() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithNonexistentAssociationIdReturnsNotFound() throws Exception {
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+
+        mockMvc.perform( patch( "/associations/9191" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"confirmed\"}" ) )
+                .andExpect( status().isNotFound() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithRemovedUpdatesAssociationStatus() throws Exception {
+        final var oldAssociationData = testDataManager.fetchAssociationDaos( "18" ).getFirst();
+
+        associationsRepository.insert( oldAssociationData );
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
+
+        mockMvc.perform( patch( "/associations/18" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+
+        final var newAssociationData = associationsRepository.findById("18").get();
+        Assertions.assertEquals( RequestBodyPut.StatusEnum.REMOVED.getValue(), newAssociationData.getStatus() );
+        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getApprovedAt() ), localDateTimeToNormalisedString( newAssociationData.getApprovedAt() ) );
+        Assertions.assertNotEquals( oldAssociationData.getRemovedAt(), newAssociationData.getRemovedAt() );
+        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
+        Assertions.assertEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
+        Assertions.assertEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithConfirmedUpdatesAssociationStatus() throws Exception {
+        final var oldAssociationData = testDataManager.fetchAssociationDaos( "18" ).getFirst();
+
+        associationsRepository.insert( oldAssociationData );
+        mockers.mockUsersServiceFetchUserDetails(  "9999" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
+
+        mockMvc.perform( patch( "/associations/18" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"confirmed\"}" ) )
+                .andExpect( status().isOk() );
+
+        final var newAssociationData = associationsRepository.findById("18").get();
+        Assertions.assertEquals( RequestBodyPut.StatusEnum.CONFIRMED.getValue(), newAssociationData.getStatus() );
+        Assertions.assertNotEquals( oldAssociationData.getApprovedAt(), newAssociationData.getApprovedAt() );
+        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getRemovedAt() ), localDateTimeToNormalisedString( newAssociationData.getRemovedAt() ) );
+        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
+        Assertions.assertEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
+        Assertions.assertEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithNullUserIdAndNonexistentUserAndConfirmedReturnsBadRequest() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "17" ) );
+        mockers.mockUsersServiceSearchUserDetailsEmptyList( "mr.blobby@nightmare.com" );
+        mockers.mockUsersServiceFetchUserDetails( "222" );
+
+        mockMvc.perform( patch( "/associations/17" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "222")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"confirmed\"}" ) )
+                .andExpect( status().isBadRequest() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithNullUserIdAndExistingUserAndRemovedUpdatesAssociationStatus() throws Exception {
+        final var associationDaos = testDataManager.fetchAssociationDaos( "1", "34" );
+        final var oldAssociationData = associationDaos.getLast();
+
+        associationsRepository.insert(associationDaos);
+        mockers.mockUsersServiceSearchUserDetails( "000" );
+        mockers.mockUsersServiceFetchUserDetails( "111" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+
+        mockMvc.perform( patch( "/associations/34" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "111")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+
+        final var newAssociationData = associationsRepository.findById("34").get();
+        Assertions.assertEquals( RequestBodyPut.StatusEnum.REMOVED.getValue(), newAssociationData.getStatus() );
+        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getApprovedAt() ), localDateTimeToNormalisedString( newAssociationData.getApprovedAt() ) );
+        Assertions.assertNotEquals( localDateTimeToNormalisedString( oldAssociationData.getRemovedAt() ), localDateTimeToNormalisedString( newAssociationData.getRemovedAt() ) );
+        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
+        Assertions.assertNotEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
+        Assertions.assertNotEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdWithNullUserIdAndNonexistentUserAndRemovedUpdatesAssociationStatus() throws Exception {
+        final var associationDaos = testDataManager.fetchAssociationDaos( "1", "34" );
+        final var oldAssociationData = associationDaos.getLast();
+
+        associationsRepository.insert(associationDaos);
+        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
+        mockers.mockUsersServiceFetchUserDetails( "111" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+
+        mockMvc.perform( patch( "/associations/34" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "111")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+
+        final var newAssociationData = associationsRepository.findById("34").get();
+        Assertions.assertEquals( RequestBodyPut.StatusEnum.REMOVED.getValue(), newAssociationData.getStatus() );
+        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getApprovedAt() ), localDateTimeToNormalisedString( newAssociationData.getApprovedAt() ) );
+        Assertions.assertNotEquals( localDateTimeToNormalisedString( oldAssociationData.getRemovedAt() ), localDateTimeToNormalisedString( newAssociationData.getRemovedAt() ) );
+        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
+        Assertions.assertEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
+        Assertions.assertEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdNotificationsWhereTargetUserIdExistsSendsNotification() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "18", "35" ) );
+        mockers.mockUsersServiceFetchUserDetails( "9999", "000" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
+
+        mockMvc.perform( patch( "/associations/35" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "9999")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdNotificationsWhereTargetUserIdDoesNotExistSendsNotification() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "1", "34" ) );
+        mockers.mockUsersServiceFetchUserDetails( "111" );
+        mockers.mockUsersServiceSearchUserDetails( "000" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+
+        setEmailProducerCountDownLatch( 1 );
+
+        mockMvc.perform( patch( "/associations/34" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "111")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+
+        latch.await( 10, TimeUnit.SECONDS );
+
+        Mockito.verify( emailProducer ).sendEmail( argThat( comparisonUtils.inviteCancelledEmailMatcher( "light.yagami@death.note", "Wayne Enterprises", "Batman" ) ), eq( INVITE_CANCELLED_MESSAGE_TYPE.getValue() ) );
+    }
+
+    @Test
+    void updateAssociationStatusForIdNotificationsWhereUserCannotBeFoundSendsNotification() throws Exception {
+        final var associationDaos = testDataManager.fetchAssociationDaos( "1", "34" );
+
+        associationsRepository.insert(associationDaos);
+        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
+        mockers.mockUsersServiceFetchUserDetails( "111" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+
+        setEmailProducerCountDownLatch( 1 );
+
+        mockMvc.perform( patch( "/associations/34" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "111")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+
+        latch.await( 10, TimeUnit.SECONDS );
+        Mockito.verify( emailProducer ).sendEmail( argThat( comparisonUtils.inviteCancelledEmailMatcher( "light.yagami@death.note", "Wayne Enterprises", "Batman" ) ), eq( INVITE_CANCELLED_MESSAGE_TYPE.getValue() ) );
+    }
+
+    @Test
+    void updateAssociationStatusForIdNotificationsUsesDisplayNamesWhenAvailable() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "3", "4" ) );
+        mockers.mockUsersServiceFetchUserDetails( "333" );
+        mockers.mockUsersServiceSearchUserDetails( "444" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+
+        mockMvc.perform( patch( "/associations/4" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "333")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"removed\"}" ) )
+                .andExpect( status().isOk() );
+    }
+
+    @Test
+    void updateAssociationToConfirmedStatusForOtherUserShouldThrow400Error() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "3", "4" ) );
+        mockers.mockUsersServiceFetchUserDetails( "333" );
+        mockers.mockUsersServiceSearchUserDetails( "444" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+
+        mockMvc.perform( patch( "/associations/4" )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "333")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"confirmed\"}" ) )
+                .andExpect( status().isBadRequest() );
+    }
+
+    @Test
+    void updateAssociationStatusForIdUserAcceptedInvitationNotificationsSendsNotification() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "2", "4", "6" ) );
+        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
+        mockers.mockUsersServiceSearchUserDetails( "666" );
+        mockers.mockUsersServiceFetchUserDetails( "222", "444", "666" );
+
+        setEmailProducerCountDownLatch( 3 );
+
+        mockMvc.perform( patch( "/associations/6"  )
+                        .header("X-Request-Id", "theId123")
+                        .header("Eric-identity", "666")
+                        .header("ERIC-Identity-Type", "oauth2")
+                        .header("ERIC-Authorised-Key-Roles", "*")
+                        .contentType( MediaType.APPLICATION_JSON )
+                        .content( "{\"status\":\"confirmed\"}" ) )
+                .andExpect( status().isOk() );
+
+        latch.await( 10, TimeUnit.SECONDS );
+
+        final var expectedBaseEmail = new InvitationAcceptedEmailBuilder()
+                .setInviterDisplayName( "the.joker@gotham.city" )
+                .setInviteeDisplayName( "homer.simpson@springfield.com" )
+                .setCompanyName( "Wayne Enterprises" );
+
+        Mockito.verify( emailProducer, times( 3 ) ).sendEmail( argThat( comparisonUtils.invitationAcceptedEmailDataMatcher( List.of( "the.joker@gotham.city", "robin@gotham.city", "homer.simpson@springfield.com" ), expectedBaseEmail ) ), eq( INVITATION_ACCEPTED_MESSAGE_TYPE.getValue() ) );
+    }
+
+    @AfterEach
+    public void after() {
+        mongoTemplate.dropCollection(AssociationDao.class);
+    }
+
+}
