+++ src/test/java/uk/gov/companieshouse/accounts/association/integration/UserCompanyAssociationsTest.java
@@ -25,16 +25,12 @@ import uk.gov.companieshouse.accounts.association.common.ComparisonUtils;
 import uk.gov.companieshouse.accounts.association.common.Mockers;
 import uk.gov.companieshouse.accounts.association.common.TestDataManager;
 import uk.gov.companieshouse.accounts.association.models.AssociationDao;
-import uk.gov.companieshouse.accounts.association.models.email.builders.InvitationAcceptedEmailBuilder;
 import uk.gov.companieshouse.accounts.association.repositories.AssociationsRepository;
 import uk.gov.companieshouse.accounts.association.service.CompanyService;
 import uk.gov.companieshouse.accounts.association.service.UsersService;
-import uk.gov.companieshouse.api.accounts.associations.model.Association;
 import uk.gov.companieshouse.api.accounts.associations.model.Association.ApprovalRouteEnum;
 import uk.gov.companieshouse.api.accounts.associations.model.Association.StatusEnum;
 import uk.gov.companieshouse.api.accounts.associations.model.AssociationsList;
-import uk.gov.companieshouse.api.accounts.associations.model.InvitationsList;
-import uk.gov.companieshouse.api.accounts.associations.model.RequestBodyPut;
 import uk.gov.companieshouse.api.accounts.associations.model.ResponseBodyPost;
 import uk.gov.companieshouse.api.handler.exception.URIValidationException;
 import uk.gov.companieshouse.email_producer.EmailProducer;
@@ -49,19 +45,13 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.times;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.companieshouse.accounts.association.common.ParsingUtils.localDateTimeToNormalisedString;
 import static uk.gov.companieshouse.accounts.association.common.ParsingUtils.parseResponseTo;
 import static uk.gov.companieshouse.accounts.association.common.ParsingUtils.reduceTimestampResolution;
 import static uk.gov.companieshouse.accounts.association.utils.MessageType.AUTH_CODE_CONFIRMATION_MESSAGE_TYPE;
-import static uk.gov.companieshouse.accounts.association.utils.MessageType.INVITATION_ACCEPTED_MESSAGE_TYPE;
-import static uk.gov.companieshouse.accounts.association.utils.MessageType.INVITATION_MESSAGE_TYPE;
-import static uk.gov.companieshouse.accounts.association.utils.MessageType.INVITE_CANCELLED_MESSAGE_TYPE;
-import static uk.gov.companieshouse.accounts.association.utils.MessageType.INVITE_MESSAGE_TYPE;
 
 @AutoConfigureMockMvc
 @SpringBootTest
@@ -125,19 +115,6 @@ class UserCompanyAssociationsTest {
                 Arguments.of( "/associations/invitations?page_index=1&items_per_page=1" )
         );
     }
-
-    @ParameterizedTest
-    @MethodSource( "withoutXRequestIdTestData" )
-    void callingEndpointsWithoutXRequestIdReturnsBadRequest( final String url ) throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( get( url )
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isBadRequest() );
-    }
-
     @Test
     void fetchAssociationsByWithoutEricIdentityReturnsForbidden() throws Exception {
         mockMvc.perform( get( "/associations" )
@@ -411,60 +388,25 @@ class UserCompanyAssociationsTest {
     }
 
     @Test
-    void getAssociationDetailsWithoutPathVariableReturnsNotFound() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( get( "/associations/" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isNotFound() );
-    }
-
-    @Test
-    void getAssociationUserDetailsWithNonexistentIdReturnsNotFound() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( get( "/associations/1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isNotFound() );
-    }
-
-
-    @Test
-    void getAssociationDetailsFetchesAssociationDetails() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "18" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
+    void fetchAssociationsByCanFetchMigratedAssociation() throws Exception {
+        associationsRepository.insert( testDataManager.fetchAssociationDaos( "MKAssociation001" ) );
+        mockers.mockUsersServiceFetchUserDetails( "MKUser001" );
+        mockers.mockCompanyServiceFetchCompanyProfile( "MKCOMP001" );
 
         final var response =
-                mockMvc.perform( get( "/associations/18" )
-                                .header("X-Request-Id", "theId123")
-                                .header("Eric-identity", "9999")
-                                .header("ERIC-Identity-Type", "oauth2")
-                                .header("ERIC-Authorised-Key-Roles", "*") )
+                mockMvc.perform( get( "/associations?status=migrated" )
+                                .header( "X-Request-Id", "theId123" )
+                                .header( "Eric-identity", "MKUser001" )
+                                .header( "ERIC-Identity-Type", "oauth2" ) )
                         .andExpect( status().isOk() );
 
-        final var association = parseResponseTo( response, Association.class );
+        final var associationsList = parseResponseTo( response, AssociationsList.class );
+        final var associations = associationsList.getItems();
+        final var associationOne = associations.getFirst();
 
-        Assertions.assertEquals( "9999", association.getUserId());
-    }
-
-    @Test
-    void addAssociationWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform(post( "/associations" )
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"111111\"}" ) )
-                .andExpect( status().isBadRequest() );
+        Assertions.assertEquals( "MKAssociation001", associationOne.getId() );
+        Assertions.assertEquals( "migrated", associationOne.getStatus().getValue() );
+        Assertions.assertEquals( "migration", associationOne.getApprovalRoute().getValue() );
     }
 
     @Test
@@ -569,7 +511,8 @@ class UserCompanyAssociationsTest {
 
         final var response = parseResponseTo( responseJson, ResponseBodyPost.class );
 
-        final var associationOptional = associationsRepository.findById( response.getAssociationId() );
+        final var associationLink = response.getAssociationLink();
+        final var associationOptional = associationsRepository.findById( associationLink.substring( associationLink.lastIndexOf( "/" )+1 ) );
         Assertions.assertTrue( associationOptional.isPresent() );
 
         final var association = associationOptional.get();
@@ -632,752 +575,6 @@ class UserCompanyAssociationsTest {
                 .andExpect( status().isCreated() );
     }
 
-    @Test
-    void updateAssociationStatusForIdWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-
-        mockMvc.perform( patch( "/associations/18" )
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithMalformedAssociationIdReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-
-        mockMvc.perform( patch( "/associations/$$$" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithoutRequestBodyReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-
-        mockMvc.perform( patch( "/associations/18" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithoutStatusReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-
-        mockMvc.perform( patch( "/associations/18" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithMalformedStatusReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-
-        mockMvc.perform( patch( "/associations/18")
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"complicated\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNonexistentAssociationIdReturnsNotFound() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-
-        mockMvc.perform( patch( "/associations/9191" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isNotFound() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithRemovedUpdatesAssociationStatus() throws Exception {
-        final var oldAssociationData = testDataManager.fetchAssociationDaos( "18" ).getFirst();
-
-        associationsRepository.insert( oldAssociationData );
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-
-        mockMvc.perform( patch( "/associations/18" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        final var newAssociationData = associationsRepository.findById("18").get();
-        Assertions.assertEquals( RequestBodyPut.StatusEnum.REMOVED.getValue(), newAssociationData.getStatus() );
-        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getApprovedAt() ), localDateTimeToNormalisedString( newAssociationData.getApprovedAt() ) );
-        Assertions.assertNotEquals( oldAssociationData.getRemovedAt(), newAssociationData.getRemovedAt() );
-        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
-        Assertions.assertEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
-        Assertions.assertEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithConfirmedUpdatesAssociationStatus() throws Exception {
-        final var oldAssociationData = testDataManager.fetchAssociationDaos( "18" ).getFirst();
-
-        associationsRepository.insert( oldAssociationData );
-        mockers.mockUsersServiceFetchUserDetails(  "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-
-        mockMvc.perform( patch( "/associations/18" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isOk() );
-
-        final var newAssociationData = associationsRepository.findById("18").get();
-        Assertions.assertEquals( RequestBodyPut.StatusEnum.CONFIRMED.getValue(), newAssociationData.getStatus() );
-        Assertions.assertNotEquals( oldAssociationData.getApprovedAt(), newAssociationData.getApprovedAt() );
-        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getRemovedAt() ), localDateTimeToNormalisedString( newAssociationData.getRemovedAt() ) );
-        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
-        Assertions.assertEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
-        Assertions.assertEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndNonexistentUserAndConfirmedReturnsBadRequest() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "17" ) );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "mr.blobby@nightmare.com" );
-        mockers.mockUsersServiceFetchUserDetails( "222" );
-
-        mockMvc.perform( patch( "/associations/17" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "222")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndExistingUserAndRemovedUpdatesAssociationStatus() throws Exception {
-        final var associationDaos = testDataManager.fetchAssociationDaos( "1", "34" );
-        final var oldAssociationData = associationDaos.getLast();
-
-        associationsRepository.insert(associationDaos);
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockUsersServiceFetchUserDetails( "111" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-
-        mockMvc.perform( patch( "/associations/34" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        final var newAssociationData = associationsRepository.findById("34").get();
-        Assertions.assertEquals( RequestBodyPut.StatusEnum.REMOVED.getValue(), newAssociationData.getStatus() );
-        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getApprovedAt() ), localDateTimeToNormalisedString( newAssociationData.getApprovedAt() ) );
-        Assertions.assertNotEquals( localDateTimeToNormalisedString( oldAssociationData.getRemovedAt() ), localDateTimeToNormalisedString( newAssociationData.getRemovedAt() ) );
-        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
-        Assertions.assertNotEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
-        Assertions.assertNotEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdWithNullUserIdAndNonexistentUserAndRemovedUpdatesAssociationStatus() throws Exception {
-        final var associationDaos = testDataManager.fetchAssociationDaos( "1", "34" );
-        final var oldAssociationData = associationDaos.getLast();
-
-        associationsRepository.insert(associationDaos);
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockUsersServiceFetchUserDetails( "111" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-
-        mockMvc.perform( patch( "/associations/34" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        final var newAssociationData = associationsRepository.findById("34").get();
-        Assertions.assertEquals( RequestBodyPut.StatusEnum.REMOVED.getValue(), newAssociationData.getStatus() );
-        Assertions.assertEquals( localDateTimeToNormalisedString( oldAssociationData.getApprovedAt() ), localDateTimeToNormalisedString( newAssociationData.getApprovedAt() ) );
-        Assertions.assertNotEquals( localDateTimeToNormalisedString( oldAssociationData.getRemovedAt() ), localDateTimeToNormalisedString( newAssociationData.getRemovedAt() ) );
-        Assertions.assertNotEquals( oldAssociationData.getEtag(), newAssociationData.getEtag() );
-        Assertions.assertEquals( oldAssociationData.getUserEmail(), newAssociationData.getUserEmail() );
-        Assertions.assertEquals( oldAssociationData.getUserId(), newAssociationData.getUserId() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsWhereTargetUserIdExistsSendsNotification() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "18", "35" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999", "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-
-        mockMvc.perform( patch( "/associations/35" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsWhereTargetUserIdDoesNotExistSendsNotification() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "1", "34" ) );
-        mockers.mockUsersServiceFetchUserDetails( "111" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-
-        setEmailProducerCountDownLatch( 1 );
-
-        mockMvc.perform( patch( "/associations/34" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-
-        Mockito.verify( emailProducer ).sendEmail( argThat( comparisonUtils.inviteCancelledEmailMatcher( "light.yagami@death.note", "Wayne Enterprises", "Batman" ) ), eq( INVITE_CANCELLED_MESSAGE_TYPE.getValue() ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsWhereUserCannotBeFoundSendsNotification() throws Exception {
-        final var associationDaos = testDataManager.fetchAssociationDaos( "1", "34" );
-
-        associationsRepository.insert(associationDaos);
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockUsersServiceFetchUserDetails( "111" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-
-        setEmailProducerCountDownLatch( 1 );
-
-        mockMvc.perform( patch( "/associations/34" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "111")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-        Mockito.verify( emailProducer ).sendEmail( argThat( comparisonUtils.inviteCancelledEmailMatcher( "light.yagami@death.note", "Wayne Enterprises", "Batman" ) ), eq( INVITE_CANCELLED_MESSAGE_TYPE.getValue() ) );
-    }
-
-    @Test
-    void updateAssociationStatusForIdNotificationsUsesDisplayNamesWhenAvailable() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "3", "4" ) );
-        mockers.mockUsersServiceFetchUserDetails( "333" );
-        mockers.mockUsersServiceSearchUserDetails( "444" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-
-        mockMvc.perform( patch( "/associations/4" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "333")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"removed\"}" ) )
-                .andExpect( status().isOk() );
-    }
-
-    @Test
-    void updateAssociationToConfirmedStatusForOtherUserShouldThrow400Error() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "3", "4" ) );
-        mockers.mockUsersServiceFetchUserDetails( "333" );
-        mockers.mockUsersServiceSearchUserDetails( "444" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-
-        mockMvc.perform( patch( "/associations/4" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "333")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void updateAssociationStatusForIdUserAcceptedInvitationNotificationsSendsNotification() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "2", "4", "6" ) );
-        mockers.mockCompanyServiceFetchCompanyProfile( "111111" );
-        mockers.mockUsersServiceSearchUserDetails( "666" );
-        mockers.mockUsersServiceFetchUserDetails( "222", "444", "666" );
-
-        setEmailProducerCountDownLatch( 3 );
-
-        mockMvc.perform( patch( "/associations/6"  )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "666")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"status\":\"confirmed\"}" ) )
-                .andExpect( status().isOk() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-
-        final var expectedBaseEmail = new InvitationAcceptedEmailBuilder()
-                        .setInviterDisplayName( "the.joker@gotham.city" )
-                        .setInviteeDisplayName( "homer.simpson@springfield.com" )
-                        .setCompanyName( "Wayne Enterprises" );
-
-        Mockito.verify( emailProducer, times( 3 ) ).sendEmail( argThat( comparisonUtils.invitationAcceptedEmailDataMatcher( List.of( "the.joker@gotham.city", "robin@gotham.city", "homer.simpson@springfield.com" ), expectedBaseEmail ) ), eq( INVITATION_ACCEPTED_MESSAGE_TYPE.getValue() ) );
-    }
-
-    @Test
-    void inviteUserWithoutXRequestIdReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWithoutEricIdentityReturnsForbidden() throws Exception {
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void inviteUserWithMalformedEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "$$$$" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "$$$$")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void inviteUserWithNonexistentEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "9191" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9191")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void inviteUserWithoutRequestBodyReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWithoutCompanyNumberReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWithMalformedCompanyNumberReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"$$$$$$\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWithNonexistentCompanyNumberReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockCompanyServiceFetchCompanyProfileNotFound( "919191" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"919191\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWithoutInviteeEmailIdReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWithMalformedInviteeEmailIdReturnsBadRequest() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\",\"invitee_email_id\":\"$$$\"}" ) )
-                .andExpect( status().isBadRequest() );
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeEmailAndCompanyNumberExistsAndInviteeUserIsFoundPerformsSwapAndUpdateOperations() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "19", "36" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}" ) )
-                .andExpect( status().isCreated() );
-
-        final var association = associationsRepository.findById( "36" ).get();
-        final var invitation = association.getInvitations().getFirst();
-
-        Assertions.assertNull( association.getUserEmail() );
-        Assertions.assertEquals( "000", association.getUserId() );
-        Assertions.assertEquals( "444444", association.getCompanyNumber() );
-        Assertions.assertEquals( StatusEnum.AWAITING_APPROVAL.getValue(), association.getStatus() );
-        Assertions.assertEquals( ApprovalRouteEnum.INVITATION.getValue(), association.getApprovalRoute() );
-        Assertions.assertNotNull( association.getApprovalExpiryAt() );
-        Assertions.assertNotNull( association.getEtag() );
-        Assertions.assertNotEquals( "qq", association.getEtag() );
-        Assertions.assertNotNull( invitation.getInvitedAt() );
-        Assertions.assertEquals( "9999", invitation.getInvitedBy() );
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeEmailAndCompanyNumberExistsAndInviteeUserIsNotFoundDoesNotPerformSwapButDoesPerformUpdateOperation() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "19", "36" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-
-        setEmailProducerCountDownLatch( 2 );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}" ) )
-                .andExpect( status().isCreated() );
-
-        final var association = associationsRepository.findById( "36" ).get();
-        final var invitation = association.getInvitations().getFirst();
-
-        Assertions.assertEquals( "light.yagami@death.note", association.getUserEmail() );
-        Assertions.assertNull( association.getUserId() );
-        Assertions.assertEquals( "444444", association.getCompanyNumber() );
-        Assertions.assertEquals( StatusEnum.AWAITING_APPROVAL.getValue(), association.getStatus() );
-        Assertions.assertEquals( ApprovalRouteEnum.INVITATION.getValue(), association.getApprovalRoute() );
-        Assertions.assertNotNull( association.getApprovalExpiryAt() );
-        Assertions.assertNotNull( association.getEtag() );
-        Assertions.assertNotEquals( "qq", association.getEtag() );
-        Assertions.assertNotNull( invitation.getInvitedAt() );
-        Assertions.assertEquals( "9999", invitation.getInvitedBy() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-        Mockito.verify( emailProducer, times( 2 ) ).sendEmail( argThat( comparisonUtils.invitationAndInviteEmailDataMatcher( "scrooge.mcduck@disney.land", "Scrooge McDuck", "light.yagami@death.note", "light.yagami@death.note", "Sainsbury's", COMPANY_INVITATIONS_URL ) ), argThat( messageType -> List.of( INVITATION_MESSAGE_TYPE.getValue(), INVITE_MESSAGE_TYPE.getValue() ).contains( messageType ) ) );
-    }
-
-    @Test
-    void inviteUserWhereInviteeUserIsFoundAndAssociationBetweenInviteeUserIdAndCompanyNumberExistsDoesNotPerformSwapButDoesPerformUpdateOperation() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "19", "36" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-
-        setEmailProducerCountDownLatch( 2 );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}" ) )
-                .andExpect( status().isCreated() );
-
-        final var association = associationsRepository.findById( "36" ).get();
-        final var invitation = association.getInvitations().getFirst();
-
-        Assertions.assertNull( association.getUserEmail() );
-        Assertions.assertEquals( "000", association.getUserId() );
-        Assertions.assertEquals( "444444", association.getCompanyNumber() );
-        Assertions.assertEquals( StatusEnum.AWAITING_APPROVAL.getValue(), association.getStatus() );
-        Assertions.assertEquals( ApprovalRouteEnum.INVITATION.getValue(), association.getApprovalRoute() );
-        Assertions.assertNotNull( association.getApprovalExpiryAt() );
-        Assertions.assertNotNull( association.getEtag() );
-        Assertions.assertNotEquals( "rr", association.getEtag() );
-        Assertions.assertNotNull( invitation.getInvitedAt() );
-        Assertions.assertEquals( "9999", invitation.getInvitedBy() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-        Mockito.verify( emailProducer, times( 2 ) ).sendEmail( argThat( comparisonUtils.invitationAndInviteEmailDataMatcher( "scrooge.mcduck@disney.land", "Scrooge McDuck", "light.yagami@death.note", "light.yagami@death.note", "Sainsbury's", COMPANY_INVITATIONS_URL ) ), argThat( messageType -> List.of( INVITATION_MESSAGE_TYPE.getValue(), INVITE_MESSAGE_TYPE.getValue() ).contains( messageType ) ) );
-    }
-
-    @Test
-    void inviteUserWhereInviteeUserIsFoundAndAssociationBetweenInviteeUserIdAndCompanyNumberDoesNotExistCreatesNewAssociation() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "19" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}" ) )
-                .andExpect( status().isCreated() );
-
-        final var association = associationsRepository.fetchAssociationForCompanyNumberAndUserId("444444", "000").get();
-        final var invitation = association.getInvitations().getFirst();
-
-        Assertions.assertNull( association.getUserEmail() );
-        Assertions.assertEquals( "000", association.getUserId() );
-        Assertions.assertEquals( "444444", association.getCompanyNumber() );
-        Assertions.assertEquals( ApprovalRouteEnum.INVITATION.getValue(), association.getApprovalRoute() );
-        Assertions.assertNotNull( association.getEtag() );
-        Assertions.assertEquals( StatusEnum.AWAITING_APPROVAL.getValue(), association.getStatus() );
-        Assertions.assertNotNull( association.getApprovalExpiryAt() );
-        Assertions.assertNotNull( invitation.getInvitedAt() );
-        Assertions.assertEquals( "9999", invitation.getInvitedBy() );
-    }
-
-    @Test
-    void inviteUserWhereAssociationBetweenInviteeUserEmailAndCompanyNumberDoesNotExistAndInviteeUserIsNotFoundCreatesNewAssociation() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "19" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetailsEmptyList( "light.yagami@death.note" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-
-        setEmailProducerCountDownLatch( 2 );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"444444\",\"invitee_email_id\":\"light.yagami@death.note\"}" ) )
-                .andExpect( status().isCreated() );
-
-        final var association = associationsRepository.fetchAssociationForCompanyNumberAndUserEmail("444444", "light.yagami@death.note").get();
-        final var invitation = association.getInvitations().getFirst();
-
-        Assertions.assertEquals( "light.yagami@death.note", association.getUserEmail() );
-        Assertions.assertNull( association.getUserId() );
-        Assertions.assertEquals( "444444", association.getCompanyNumber() );
-        Assertions.assertEquals( ApprovalRouteEnum.INVITATION.getValue(), association.getApprovalRoute() );
-        Assertions.assertNotNull( association.getEtag() );
-        Assertions.assertEquals( StatusEnum.AWAITING_APPROVAL.getValue(), association.getStatus() );
-        Assertions.assertNotNull( association.getApprovalExpiryAt() );
-        Assertions.assertNotNull( invitation.getInvitedAt() );
-        Assertions.assertEquals( "9999", invitation.getInvitedBy() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-        Mockito.verify( emailProducer, times( 2 ) ).sendEmail( argThat( comparisonUtils.invitationAndInviteEmailDataMatcher( "scrooge.mcduck@disney.land", "Scrooge McDuck", "light.yagami@death.note", "light.yagami@death.note", "Sainsbury's", COMPANY_INVITATIONS_URL ) ), argThat( messageType -> List.of( INVITATION_MESSAGE_TYPE.getValue(), INVITE_MESSAGE_TYPE.getValue() ).contains( messageType ) ) );
-    }
-
-    @Test
-    void inviteUserWhereInviteeUserIsFoundAndAssociationBetweenInviteeUserIdAndCompanyNumberExistsDoesNotPerformSwapButThrowsBadRequest() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "18", "35" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetails( "000" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "333333" );
-
-        mockMvc.perform( post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"333333\",\"invitee_email_id\":\"light.yagami@death.note\"}" ) )
-                .andExpect( status().isBadRequest() ).andReturn();
-    }
-
-    @Test
-    void inviteUserWithUserThatHasDisplayNameUsesDisplayName()  throws Exception {
-
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "19" ) );
-        mockers.mockUsersServiceFetchUserDetails( "9999" );
-        mockers.mockUsersServiceSearchUserDetails( "111" );
-        mockers.mockCompanyServiceFetchCompanyProfile( "444444" );
-
-        setEmailProducerCountDownLatch( 2 );
-
-        mockMvc.perform(post( "/associations/invitations" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9999")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*")
-                        .contentType( MediaType.APPLICATION_JSON )
-                        .content( "{\"company_number\":\"444444\",\"invitee_email_id\":\"bruce.wayne@gotham.city\"}" ) )
-                .andExpect( status().isCreated() );
-
-        latch.await( 10, TimeUnit.SECONDS );
-        Mockito.verify( emailProducer, times( 2 ) ).sendEmail( argThat( comparisonUtils.invitationAndInviteEmailDataMatcher( "scrooge.mcduck@disney.land", "Scrooge McDuck", "bruce.wayne@gotham.city", "Batman", "Sainsbury's", COMPANY_INVITATIONS_URL ) ), argThat( messageType -> List.of( INVITATION_MESSAGE_TYPE.getValue(), INVITE_MESSAGE_TYPE.getValue() ).contains( messageType ) ) );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithoutEricIdentityReturnsForbidden() throws Exception {
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithMalformedEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "$$$$" );
-
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "$$$$")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithNonexistentEricIdentityReturnsForbidden() throws Exception {
-        mockers.mockUsersServiceFetchUserDetailsNotFound( "9191" );
-
-        mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                        .header("X-Request-Id", "theId123")
-                        .header("Eric-identity", "9191")
-                        .header("ERIC-Identity-Type", "oauth2")
-                        .header("ERIC-Authorised-Key-Roles", "*") )
-                .andExpect( status().isForbidden() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserRetrievesActiveInvitationsInCorrectOrderAndPaginatesCorrectly() throws Exception {
-        associationsRepository.insert( testDataManager.fetchAssociationDaos( "37", "38" ) );
-        mockers.mockUsersServiceFetchUserDetails(  "000", "444" );
-
-        final var response =
-                mockMvc.perform( get( "/associations/invitations?page_index=0&items_per_page=1" )
-                                .header("X-Request-Id", "theId123")
-                                .header("Eric-identity", "000")
-                                .header("ERIC-Identity-Type", "oauth2")
-                                .header("ERIC-Authorised-Key-Roles", "*") )
-                        .andExpect( status().isOk() );
-
-        final var invitations = parseResponseTo( response, InvitationsList.class ).getItems();
-        final var invitation = invitations.getFirst();
-
-        Assertions.assertEquals( 1, invitations.size() );
-        Assertions.assertEquals( "robin@gotham.city", invitation.getInvitedBy() );
-        Assertions.assertNotNull( invitation.getInvitedAt() );
-        Assertions.assertEquals( "37", invitation.getAssociationId() );
-        Assertions.assertTrue( invitation.getIsActive() );
-    }
-
-    @Test
-    void fetchActiveInvitationsForUserWithoutActiveInvitationsReturnsEmptyList() throws Exception {
-        mockers.mockUsersServiceFetchUserDetails(  "111" );
-
-        final var response =
-                mockMvc.perform( get( "/associations/invitations?page_index=1&items_per_page=1" )
-                                .header("X-Request-Id", "theId123")
-                                .header("Eric-identity", "111")
-                                .header("ERIC-Identity-Type", "oauth2")
-                                .header("ERIC-Authorised-Key-Roles", "*") )
-                        .andExpect( status().isOk() );
-
-        final var invitations = parseResponseTo( response, InvitationsList.class );
-        Assertions.assertTrue( invitations.getItems().isEmpty() );
-    }
 
     @AfterEach
     public void after() {
